+++ Assets/StarterKit/Libraries/Markdig/MarkdigExtensions/Renderers/UxmlEmphasisInlineRenderer.cs
@@ -1,0 +1,43 @@
+using Markdig.Syntax.Inlines;
+using System.Diagnostics;
+using Markdig.Renderers.Uxml;
+using Debug = UnityEngine.Debug;
+
+namespace Markdig.Renderers.Html {
+
+/// <summary>
+/// A HTML renderer for an <see cref="EmphasisInline"/>.
+/// </summary>
+/// <seealso cref="HtmlObjectRenderer{EmphasisInline}" />
+public class UxmlEmphasisInlineRenderer : UxmlObjectRenderer<EmphasisInline> {
+
+    private MarkdownSettings renderSettings;
+
+    public UxmlEmphasisInlineRenderer(MarkdownSettings renderSettings) {
+        this.renderSettings = renderSettings;
+    }
+
+    protected override void Write(UxmlRenderer renderer, EmphasisInline obj) {
+        string opener = (obj.DelimiterChar switch {
+            '*' => "<b>",
+            '_' => "<i>",
+            '~' => "<s>",
+            _ => ""
+        });
+        string closer = (obj.DelimiterChar switch {
+            '*' => "</b>",
+            '_' => "</i>",
+            '~' => "</s>",
+            _ => ""
+        });
+
+
+        renderer.WriteRaw(opener);
+
+        renderer.WriteChildren(obj);
+        
+        renderer.WriteRaw(closer);
+
+    }
+}
+}