+++ Assets/StarterKit/CHANGELOG.md
@@ -1,9 +1,34 @@
 # Changelog
 
+## [0.4.0]
+
+### Cheat Sheet
+- Added Math Section
+- Added 2D Section
+- Added Physics Section
+- Added Audio Section
+- Added UI Section
+- Added Managing Data Section
+- Added Editor Scripting Section
+- Added Multiplayer Section
+- Re-Ordered Sections
+- Added more headers for future sections
+
+### Editor Scripts
+- Added Divider Field Attribute
+- Added Button Field Attribute
+
+### Scriptable Objects
+- Added Checklist (Assets > Create > Noodle Kit > Checklist)
+
+### Utils
+- Added NoodleUtils Class
+- Added FormatCamelCase() function to utils
+
 ## [0.3.0]
 
 ### Cheat Sheet
-- Added Input sections
+- Added Input section
 - Added Math section
 - Added 2D section